requirements.txt
streamlit==1.47.1
pymongo==4.6.1
python-dotenv==1.0.0
openai==1.47.1
requests==2.31.0
httpx==0.27.0
pandas==2.2.2
